[DAD-1137] 로그아웃, 탈퇴하기 버튼 컴포넌트 분리
